Clean up Lambda dedup, dead stubs, logging, and terraform refs
Extract shared get_device_id() and send_sidewalk_msg() into
sidewalk_utils.py to eliminate duplication between ota_sender and
charge_scheduler Lambdas. Delete 5 dead platform-side stub files
superseded by split-image app_evse/. Add warning logs to
charge_control silent error returns. Replace hardcoded "ota-sender"
strings in terraform with resource references.
